coldfront/plugins/freeipa/requirements.txt
ipaclient==4.7.2
gssapi==1.5.1
ldap3==2.6
dbus-python==1.2.8
